+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - Added RuboCop to the default rake task
 - Added RuboCop for rake
 - Added RuboCop for RSpec
+- Added module to provide helper methods for working with Object Paths
+  - Added class & instance methods for creating Object Paths
+  - Added method to directly resolve Object Paths against the current instance
 
 # 1.0.2 / 2024-09-18